Proposed Ending Event output
This commit adds a second output to the Events Node that fires at the end of the event. To allow for the event to be changed at anytime in the calendar, the uid is constructed here from the ev.created parameter instead of started. The words start and end are appended to the end of the uid. Thus each calendar event has a pair of cron jobs. Lastly, the else if block that checks if a started job already exists updates that job with the new run time.